Support building the toolkit as a standalone extension (#4208)
