+++ plugins/amazonq/codetransform/jetbrains-community/build.gradle.kts
@@ -11,6 +11,10 @@ intellijToolkit {
     ideFlavor.set(IdeFlavor.IC)
 }
 
+intellij {
+    plugins.add(project(":plugin-core"))
+}
+
 dependencies {
     implementation(project(":plugin-amazonq:shared:jetbrains-community"))
     // hack because transform has a chat entrypoint
@@ -19,4 +23,13 @@ dependencies {
     compileOnly(project(":plugin-core:jetbrains-community"))
 
     testImplementation(testFixtures(project(":plugin-core:jetbrains-community")))
+}
+
+// hack because our test structure currently doesn't make complete sense
+tasks.prepareTestingSandbox {
+    val pluginXmlJar = project(":plugin-amazonq").tasks.jar
+
+    dependsOn(pluginXmlJar)
+    intoChild(pluginName.map { "$it/lib" })
+        .from(pluginXmlJar)
 }
